requirements.tf1.txt
absl-py==0.7.1
appdirs==1.4.3
asciimatics==1.10.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
configobj==5.0.6
configparser==3.7.3
contextlib2==0.5.5
decorator==4.3.2
defusedxml==0.5.0
distro==1.4.0
dvc==0.30.1
entrypoints==0.3
Flask==1.0.3
future==0.17.1
gast==0.2.2
gitdb2==2.0.5
GitPython==2.1.11
google-pasta==0.1.4
grandalf==0.6
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonpath-rw==1.4.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
nanotime==0.5.2
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.6
numpy==1.16.4
opencv-python==4.1.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pickleshare==0.7.5
Pillow==5.4.1
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
pyasn1==0.4.5
pyfiglet==0.8.post1
Pygments==2.3.1
pyparsing==2.3.1
pypiwin32==223
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.3
pywin32==224
pywinpty==0.5.5
PyYAML==3.13
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
schema==0.7.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
smmap2==2.0.5
tb-nightly==1.14.0a20190301
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
tornado==6.0.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
zc.lockfile==1.4
